dataviz-backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
psycopg2-binary
sqlalchemy
python-dotenv
requests
jieba
openai
python-multipart